# Conflicts:
#	app/services/admin_management/kyc/check_details.rb
#	set_env_vars.sh

+++ app/controllers/web/admin/base_controller.rb
@@ -10,14 +10,13 @@ class Web::Admin::BaseController < Web::WebController
   # * Date: 10/10/2017
   # * Reviewed By: Sunil Khedar
   #
-  def authenticate_request(is_super_admin_role=false)
+  def authenticate_request(options={is_super_admin_role: false, validate_terms_of_use: true})
 
     service_response = AdminManagement::VerifyCookie::DoubleAuth.new(
         cookie_value: cookies[GlobalConstant::Cookie.admin_cookie_name.to_sym],
         browser_user_agent: http_user_agent,
-        is_super_admin_role: is_super_admin_role
+        options: options
     ).perform
-
     if service_response.success?
       # Update Cookie, if required
       extended_cookie_value = service_response.data[:extended_cookie_value]
@@ -33,7 +32,9 @@ class Web::Admin::BaseController < Web::WebController
       # Remove sensitive data
       service_response.data = {}
     else
-      delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+      if service_response.http_code == GlobalConstant::ErrorCode.unauthorized_access
+        delete_cookie(GlobalConstant::Cookie.admin_cookie_name)
+      end
       render_api_response(service_response)
     end
